split_train_val function also moved to continuumDataset file
